% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{state number, using fars state numbering convention}

\item{year}{the years of interest with 4 digits.}
}
\value{
This function returns a tibble with the number of incidents broken
 down by month (rows) and year (column(s)).
}
\description{
Given a state number and year, this function returns a plots the incidents
on a map of the state by latitude and longitude coordinates. If the input
includes an invalid state number or year, an error message will occur. If
there are are no incidents to plot, a message will be displayed.
Functions from other packages that will be needed are: dplyr::filter,
maps::map, and graphics::points
}
\examples{
fars_map_state(1,2013)

}
