% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\source{
\url{http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
Functions from other packages that will be needed are: readr::read_csv and dplyr::tbl_df
}
\usage{
fars_read(filename, package_file = TRUE)
}
\arguments{
\item{filename}{the name of the file which the data are to be read from.}

\item{package_file}{if TRUE, this function looks for files in the package
directory inst/extdata. Otherwise, the file needs to be in the working directory.}
}
\value{
This function returns a tbl_df object
}
\description{
tbl_df object. If the filename does not exist, error message will be displayed.
Data should be sourced from the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System, which is a nationwide census providing
the American public yearly data regarding fatal injuries suffered in motor
vehicle traffic crashes.
}
\examples{
fars_read(make_filename(2013))

}
