% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{the years of interest with 4 digits. This can be a vector
of years including the}
}
\value{
This function returns a list of tibbles, one tibble per year. The tibble
includes the MONTH and year columns of the fars data
}
\description{
Given a year or vector of years, this function returns a list of tibbles with
the MONTH and year columns of the fars data. If the input includes an invalid
year, that particular tibble will be \code{NULL} and a warning message will be
displayed.
Functions from other packages that will be needed are: dplyr::mutate and dplyr::select
}
\examples{
fars_read_years(c(2013,2014,2015))

}
