% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{the years of interest with 4 digits. This can be a vector}
}
\value{
This function returns a tibble with the number of incidents broken
 down by month (rows) and year (column(s)).
}
\description{
Given a year or vector of years, this function returns a tibble with the number
of incidents broken down by month and year. If the input includes an invalid
year, an error message will occur.
Functions from other packages that will be needed are: dplyr::bind_rows,
dplyr::group_by, dplyr::summarize, and tidyr::spread
}
\examples{
fars_summarize_years(c(2013,2014,2015))

}
